+++ build-tools/config/webpack/webpack.conf.prod.js
@@ -1,0 +1,1 @@
+module.exports = require('./webpack.conf.base')(false);
