+++ build-tools/config/webpack/webpack.helpers.js
@@ -1,0 +1,11 @@
+exports.compose = function(funcs) {
+  if (funcs.length === 0) {
+    return arg => arg
+  }
+
+  if (funcs.length === 1) {
+    return funcs[0]
+  }
+
+  return funcs.reduce((a, b) => (...args) => a(b(...args)));
+};
