+++ build-tools/config/webpack/webpack.partial.conf.entry.js
@@ -1,0 +1,9 @@
+module.exports = (config, isDevelopment) => webpackConfig => ({
+  ...webpackConfig,
+  entry: {
+    app: [
+      './src/polyfill/polyfill.js',
+      './src/bootstrap.js'
+    ],
+  },
+});
