+++ build-tools/config/webpack/webpack.partial.conf.node.js
@@ -1,0 +1,15 @@
+module.exports = (config, isDevelopment) => webpackConfig => ({
+  ...webpackConfig,
+  node: {
+    // prevent webpack from injecting useless setImmediate polyfill because Vue
+    // source contains it (although only uses it if it's native).
+    setImmediate: false,
+    // prevent webpack from injecting mocks to Node native modules
+    // that does not make sense for the client
+    dgram: 'empty',
+    fs: 'empty',
+    net: 'empty',
+    tls: 'empty',
+    child_process: 'empty',
+  },
+});
