+++ build-tools/config/webpack/webpack.partial.conf.optimization.js
@@ -1,0 +1,20 @@
+module.exports = (config, isDevelopment) => webpackConfig => ({
+  ...webpackConfig,
+  optimization: isDevelopment ? {
+    noEmitOnErrors: true,
+  } : {
+    concatenateModules: true,
+    minimize: true,
+    splitChunks: {
+      chunks: 'all',
+      cacheGroups: {
+        commons: {
+          test: /[\\/]node_modules[\\/]/,
+          name: 'vendors',
+          chunks: 'all',
+        },
+      },
+    },
+    runtimeChunk: false,
+  },
+});
